refactor: ongoing work on resolvedcapabilities insert and permits fn impl
So I was going to reimplement the final `ResolvedCapabilities` type as a `HashMap<ResolvedResource, HashMap<Ability, Caveats>>`. This was supposed to be a canonical representation that would be easy to query for permission. However, after spending days trying to figure it out, I realized that making it canonical was non-trivial and I'm basically optimizing prematurely.

For example, Given the get_permit_levels implementation, if you insert `zerodb:// x crud/delete x _` and then try to insert `zerodb://public x crud/* x _`, you get a `Permitted x ReversePermitted x _`  which would have to be inserted as its own separate thing. Now if you try and insert `zerodb:// x crud/* x _` you get `Same x ReversePermitted x _` or `ReversePermitted x Same x _` which depending on which of the existing two entries is tested first. And because only one of them is recognized and not both, you get left with a non-canonical ResolvedCapabilities collection defeating the purpose of fast permit check. So I'm ggoing back to the HashSet<ResolveCapabilityTuple> implementation. No more optimization that waste days and it also simplifies implementation which is great for security.

Remove CapabilitiesDefinition because it was getting complex. I had to think of a way to convert Caveats to CaveatsDefinition JTD and it wasn't going to be reusable as canonical ResolvedCapabilities. I figured definition is not necessary as validation would be done at runtime anyway by verifier.